Remove some set/restore C locale calls
In a previous commit, I put some XmTextInsert calls inside
set_to_C_locale/restore_from_C_locale calls.

The intent of this was to prevent warnings about invalid character
260 (degree sign) when rendering the station info dialogs (scrollable
text that includes bearing-from-my-station and current course info of
a station).

Unfortunately, these made the degree symbol render oddly when the
locale was en_US.UTF-8 --- it would render as an A with a circumflex
followed by a degree symbol.

Oddly, when the locale was *already* C (i.e. LANG=C set before running
Xastir), these rendered fine.

Without these pairs of calls, and with LANG=en_US.UTF-8, the degree
symbol doesn't render at all, but doesn't spew the warning.

I have opted to remove them, returning this part of the code to what
it was this afternoon, because that oddly-rendered degree symbol looks
worse to me than the unrendered degree symbol.

But it's still wrong.  Something about how we're draing text
WPUPSTI022 and the per-packet station course angle is goofy, and only
works correctly when the locale under which the scrollable text widget
is created is C, not the locale active while drawing into the widget.

Since setting the locale to C at the time the widget is created seems
wrong (the locale may be a language other than English, and this may
break the rendering of the dialog), I am at a loss for how to proceed
here.  So rather than leave it half-baked, I'm removing the code I
just added.

I am leaving the other code I changed, in draw_symbol, because that
works just fine no matter what the locale is, and fixes the annoying
warning.