Move bstring to subdirectory
This will keep the "Better String Library" in a separate subdirectory
making it easier to keep it separate from Xastir main as well as
easier to update the library in the future. We have bstrlib and
bstraux linked in so have access to both the core and the auxiliary
functions of bstring.